pyproject.toml
[tool.poetry]
name = "discordwrap"
version = "0.4.0"
description = ""
authors = ["josh <josh@brunuslabs.com>"]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
python-dotenv = "1.0.0"
python-semantic-release = "^7.31.2"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "discordwrap/__init__.py:__version__"
]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
version_pattern = [
    "README.md:rev: v{version}"
]
major_on_zero = false
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "pip install poetry && poetry build"

